+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Changed
+
+- Remove SOPS_AGE_KEY_FILE from default environment variables.
+
 ## [1.3.0] - 2025-10-21
 
 ### Added