+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+## [1.4.0] - 2026-04-06
+
 ### Added
 
 - Support for per-project environment variable configuration for SOPS.
@@ -44,7 +46,8 @@
 
 - Initial implementation of the plugin.
 
-[Unreleased]: https://github.com/Blarc/sops-intellij-plugin/compare/v1.3.0...HEAD
+[Unreleased]: https://github.com/Blarc/sops-intellij-plugin/compare/v1.4.0...HEAD
+[1.4.0]: https://github.com/Blarc/sops-intellij-plugin/compare/v1.3.0...v1.4.0
 [1.3.0]: https://github.com/Blarc/sops-intellij-plugin/compare/v1.2.0...v1.3.0
 [1.2.0]: https://github.com/Blarc/sops-intellij-plugin/compare/v1.1.1...v1.2.0
 [1.1.1]: https://github.com/Blarc/sops-intellij-plugin/compare/v1.1.0...v1.1.1